Refactor:#82 - move for schema validation.ts to auth.ts
